migrate testing tool to jest
