+++ packages/server/package.json
@@ -9,12 +9,12 @@
   "scripts": {
     "compile": "tsc -p .",
     "watch": "tsc --watch",
-    "test": "tsc && ava dist/test",
-    "prepublish": "yarn run compile"
+    "test": "jest"
   },
   "dependencies": {
     "@joe-re/node-sql-parser": "^0.0.1",
     "@types/pg": "^7.4.10",
+    "jest": "^26.0.1",
     "log4js": "^2.5.3",
     "mysql": "^2.15.0",
     "pg": "^7.4.3",
@@ -22,9 +22,11 @@
     "yargs": "^12.0.1"
   },
   "devDependencies": {
+    "@types/jest": "^25.2.2",
     "@types/mysql": "^2.15.5",
     "@types/node": "^8.10.0",
-    "ava": "^0.25.0",
-    "typescript": "^2.9.1"
+    "jest": "^26.0.1",
+    "ts-jest": "^26.0.0",
+    "typescript": "^3.9.2"
   }
 }
